Add try/catch to log issues with ENR records
